Refactor of "resolvers" folder to "Resolvers"
