+++ src/index.js
@@ -1,16 +1,16 @@
 const { ApolloServer, gql, makeExecutableSchema } = require("apollo-server");
 const { Prisma } = require("prisma-binding");
 const {EmailAddress, PostalCode} =  require("@okgrow/graphql-scalars");
-const Query = require("./resolvers/Query");
-const {modifyProfile, createTeam, modifyTeam, deleteTeam, modifyApproval} = require("./resolvers/Mutations");
-const {PhoneNumber} = require("./resolvers/Scalars");
+const Query = require("./Resolvers/Query");
+const {modifyProfile, createTeam, modifyTeam, deleteTeam, modifyApproval} = require("./Resolvers/Mutations");
+const {PhoneNumber} = require("./Resolvers/Scalars");
 const config = require("./config");
 const AuthDirectives = require("./Auth/Directives");
 const fs = require("fs");
 const { connectMessageQueueListener } = require("./Service_Mesh/listener_connector");
 const { connectMessageQueuePublisher } = require("./Service_Mesh/publisher_connector");
 const introspect = require("./Auth/introspection");
-const { getDefaults } = require("./resolvers/helper/default_setup");
+const { getDefaults } = require("./Resolvers/helper/default_setup");
 const { applyMiddleware } = require("graphql-middleware");
 const { approvalRequired } = require("./Middleware/approvalCreation");
 const {allowedToModifyProfile, allowedToModifyApproval, mustbeAuthenticated} = require("./Middleware/authMiddleware");
